requirements.txt
pyspark>=3.1.2
pyyaml>=6.0
flask